more debugging

+++ setup_ci.sh
@@ -7,8 +7,18 @@ env
 
 echo $PATH
 
-curl --proto '=https' --tlsv1.2 -sSf -L https://install.determinate.systems/nix | sh -s -- install linux \
-  --init none \
-  --extra-conf "sandbox = false" \
-  --no-start-daemon \
-  --no-confirm
+sudo env
+
+sudo which nix-build
+
+if [ -f /nix/receipt.json ]; then
+  echo "Nix is already installed"
+else
+  echo "Installing Nix"
+
+  curl --proto '=https' --tlsv1.2 -sSf -L https://install.determinate.systems/nix | sh -s -- install linux \
+    --init none \
+    --extra-conf "sandbox = false" \
+    --no-start-daemon \
+    --no-confirm
+fi
